requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
openai==1.42.0
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
